Failing test to illustrate issue with dup packets
We ran a playback exercise using a miniseed dataset that had some
duplicate packets, which caused obspy's data loader to generate many
traces per channel. The current w-phase code interprets this as gappy
data and throws it all out.

Required a small refactor to expose the gap removal for testing.