Fix validation test
IRIS data availability must have changed, since the results of this
inversion test case have changed. I've modified the tests so that the
validation is performed using a fixed input dataset.

Left a test using IRIS FDSN in there, but it just checks that it
converges rather than validating the results.