Clean up logging, refactor result emails.
Moved the CLI app to a module so that it can be extended with custom
behaviour.

Moved email body generation out of wphase.app into a separate method
that can be overridden for custom email messages.

Replaced print-logging with a python logging.Logger, and redirected the
output of this logger to the seiscomp logs when using the CLI app.