Consolidate W-Phase result values with Pydantic
We previously had results being passed around in three different formats:

- Big old tuples full of heavyweight data for e.g. plotting, which I
  recently replaced with the dataclass-like OL1, OL2, OL3 classes
- The output dictionary, which is what gets written to
  wphase_output.json at the end
- The FMItem class, which parsed an output dictionary into yet another
  custom object used by wphase.seiscomp.

I've refactored the code to replace all three of these with a single
pydantic schema. The schema follows the exact structure of
wphase_output.json with a few extra fields (for the heavyweight data)
attached, but these fields are not serialized; so we achieve the same
output json just by calling pydantic's .json() method.