Improve channel selection
Previously we were taking a very coarse approach to channel selection:
We just took anything with location code 00/--/blank, and accepted only
data with band B or L. This meant that e.g. NZ data was excluded from
our calculations (both because it was H-band and because it was location
10).

I've replaced this with a smarter selection routine that will do its
best to include the one best channel for each station.